+++ src/ManagedShell.WindowsTray/TrayService.cs
@@ -239,8 +239,8 @@ namespace ManagedShell.WindowsTray
 
                     if ((wndPos.flags & SetWindowPosFlags.SWP_SHOWWINDOW) != 0)
                     {
-                        SetWindowLongPtr(HwndTray, (int)WindowLongFlags.GWL_STYLE,
-                            (IntPtr)((int)GetWindowLongPtr(HwndTray, (int)WindowLongFlags.GWL_STYLE) &
+                        SetWindowLongPtr(HwndTray, WindowLongFlags.GWL_STYLE,
+                            (IntPtr)((int)GetWindowLongPtr(HwndTray, WindowLongFlags.GWL_STYLE) &
                             ~(int)WindowStyles.WS_VISIBLE));
 
                         ShellLogger.Debug($"TrayService: {TrayWndClass} became visible; hiding");
